Separate ranking and filtering regex logic.
Basically, we use the faster regex (match as fast as we can) to filter
and the better regex (optimize for clusters of strings) for ranking.
True, we have to run both on all matches, but in any meaningful example,
this'll be preferable to running the slow regex on everything.

Obviates #16 for now.