Allow directory switching via input string.
If you start typing a relative or absolute path, we'll start searching
in that directory.  That is, if you type ~/..., we'll look your home
directory.  /tmp/... means we'll search /tmp.  It's a bit tricky to get
the caching to work, and the filename search thread needs to be in sync
with the candidate filename thread, but I think I got it.

Fixes #5.