In progress.
I can read things in and dump things out to the command line, but the
string parsing is whack.  For example, if I have a quoted string already
on the command line that doesn't need to be quoted, it strips the
quotes, which throws off where I should be setting READLINE_POINT.  It
needs a bit more love before being ready for prime time.

This is for issue #8.